+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Enhancements to GandiV5::Domain
   * .create now returns created domain (except in dry-run mode)
+  * .tlds moved to GandiV5::Domain::TLD.list
+  * .tld(name) moved to GandiV5::Domain::TLD.fetch(name)
 * Enhancements to GandiV5::Email::Mailbox
   * .create now returns created mailbox
   * .create now checks for available slots and a valid type has been passed